requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
psycopg2-binary>=2.8
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
whitenoise==5.2.0